[댓글 기능] 올레 미션 제출합니다. (#113)
* refactor: use httpSession directly instead using request

* refactor: use optional at findByEmail

* refactor: move function of matching password to User from LoginService

* refactor: use RedirectionView object for redirection

* refactor: use User parameter at delete and save method

* refactor: delete loginDTO and use UserDTO instead

* refactor: delete update method that use jpql query in UserRepository

use update function of User and add Transactiional annotation instead

* refactor: delete update method that use jpql query in ArticleRepository

use update function of Article and add Transactiional annotation instead

* feat: add reference of article and user entity to each other

* feat: add new ArrayList at articles field in User class

* feat: add column length constraint at User

* docs: write todo file by content of 3rd mission

* docs: add 3rd mission required to readme file

* docs: add check markdown to todo file

* fix: bug that password checking logic does not work

change password to confirmpasssword at signup.html

* feat: comment and comment repository

add test case of comment repository's CRUD method

* feat: create comment and show comments

* feat: delete comment and show comment count at index page

* refactor: change request mapping parameter

remove duplicated url

* bug: comment editting window is not shown at article page

* fix: previous bug fixed

405 not found bug remained

* fix: 405 not found of put method(comment edit) bug fixed

* feat: alert error message at article page

* feat: block unauthorized author of article

* refactor: use equals method at comparing user instead using '=='

* refactor: add cascade option at @ManyToOne at Article

* refactor: put user object into article-edit html at createform method in ArticleController class

* refactor: change method name of controller

* refactor: delete ArticleController test file and CommentController test file

* refactor: add ArticleControllerTest

still some of tests fails

* refactor: tests of ArticleControllerTest all success

* refactor: use nested tests at ArticleControllerTest

still editting feature test fails

* refactor: check whether user is author when user tries to edit

test all success

* refactor: delete setAuthor in Article class and add user parameter at ArticleDTO.toDomain method

* refactor: add isAuthorOf method at user class

checks whether user is author of article parameter

* refactor: delete name variable of PathVariable annotation

* refactor: add user in model, at get request mapping of article edit page

* refactor: delete setter method at comment and add parameters to CommentDTO.toDomain method

* refactor: move finding article method code right above of article using code

* refactor: add error logger at exception handling method

* refactor: add CommentController tests at CommentControllerTest file

* refactor: set message string field at NotFoundCommentException and UnAuthenticatedUserException

* refactor: add code that checks whether user exists at LoginService.getLoginUser method

set message string field at NotFoundUserException

* refactor: delete articleRepository tests and commentRepository tests

* refactor: add lazy fetching at ManyToOne annotation at entity classes

delete private access of noArgsConstructor, because it gives you error when you do lazy fetching

+++ src/main/java/techcourse/myblog/controller/ArticleController.java
@@ -1,55 +1,128 @@
 package techcourse.myblog.controller;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Controller;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
+import org.springframework.web.servlet.view.RedirectView;
 import techcourse.myblog.domain.Article;
+import techcourse.myblog.domain.User;
+import techcourse.myblog.exception.NotFoundArticleException;
+import techcourse.myblog.exception.UnauthenticatedUserException;
 import techcourse.myblog.repository.ArticleRepository;
 import techcourse.myblog.service.dto.ArticleDTO;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
 
 @Controller
 @RequestMapping("/articles")
 public class ArticleController {
     private final ArticleRepository articleRepository;
+    private static final Logger log = LoggerFactory.getLogger(ArticleController.class);
+
 
     public ArticleController(ArticleRepository articleRepository) {
         this.articleRepository = articleRepository;
     }
 
     @GetMapping("/{articleId}")
-    public String showArticle(@PathVariable("articleId") Long articleId, Model model) {
-        model.addAttribute("article", articleRepository.findById(articleId).get());
-        model.addAttribute("articleId", articleId);
+    public String show(@PathVariable Long articleId, HttpSession session, Model model) {
+        Article article = articleRepository
+                .findById(articleId)
+                .orElseThrow(NotFoundArticleException::new);
+
+        model.addAttribute("article", article);
+        model.addAttribute("comments", article.getComments());
+
+        User user = (User) session.getAttribute("user");
+        if (user != null) {
+            model.addAttribute("user", user);
+        }
+
         return "article";
     }
 
     @GetMapping("/new")
-    public String articleCreateForm() {
+    public String createForm(HttpSession session, Model model) {
+        User user = (User) session.getAttribute("user");
+        if (user == null) {
+            throw new UnauthenticatedUserException();
+        }
+        model.addAttribute("user", user);
         return "article-edit";
     }
 
     @PostMapping("/write")
-    public String saveArticle(ArticleDTO articleDTO) {
-        Article article = articleRepository.save(articleDTO.toDomain());
+    public String create(ArticleDTO articleDTO, HttpSession session) {
+        User user = (User) session.getAttribute("user");
+        if (user == null) {
+            throw new UnauthenticatedUserException();
+        }
+
+        Article article = articleDTO.toDomain(user);
+
+        articleRepository.save(article);
         return "redirect:/articles/" + article.getId();
     }
 
     @GetMapping("/{articleId}/edit")
-    public String articleUpdateForm(@PathVariable("articleId") Long articleId, Model model) {
-        model.addAttribute("article", articleRepository.findById(articleId).get());
+    public Object updateForm(@PathVariable Long articleId, HttpSession session, Model model) {
+        Article savedArticle = articleRepository
+                .findById(articleId)
+                .orElseThrow(NotFoundArticleException::new);
+
+        checkAuthor(session, savedArticle);
+
+        model.addAttribute("user", session.getAttribute("user"));
+        model.addAttribute("article", savedArticle);
         model.addAttribute("articleId", articleId);
         return "article-edit";
     }
 
+    @Transactional
     @PutMapping("/{articleId}")
-    public String updateArticle(@PathVariable("articleId") Long articleId, ArticleDTO article) {
-        articleRepository.updateArticleById(article.getTitle(), article.getCoverUrl(), article.getContents(), articleId);
+    public String update(@PathVariable Long articleId, HttpSession session, ArticleDTO articleDTO) {
+        Article savedArticle = articleRepository
+                .findById(articleId)
+                .orElseThrow(NotFoundArticleException::new);
+
+        checkAuthor(session, savedArticle);
+        savedArticle.update(articleDTO.toDomain(savedArticle.getAuthor()));
+
         return "redirect:/articles/" + articleId;
     }
 
+    private void checkAuthor(HttpSession session, Article savedArticle) {
+        User user = (User) session.getAttribute("user");
+        if (!user.equals(savedArticle.getAuthor())) {
+            throw new UnauthenticatedUserException();
+        }
+    }
+
     @DeleteMapping("/{articleId}")
-    public String deleteArticle(@PathVariable("articleId") Long articleId) {
+    public String delete(@PathVariable Long articleId, HttpSession session) {
+        Article savedArticle = articleRepository
+                .findById(articleId)
+                .orElseThrow(NotFoundArticleException::new);
+
+        User user = (User) session.getAttribute("user");
+        if (!user.isAuthorOf(savedArticle)) {
+            throw new UnauthenticatedUserException();
+        }
+
         articleRepository.deleteById(articleId);
         return "redirect:/";
+    }
+
+    @ExceptionHandler(RuntimeException.class)
+    public RedirectView exceptionHandler(RuntimeException exception, RedirectAttributes redirectAttributes, HttpServletRequest request) {
+        redirectAttributes.addFlashAttribute("articleError", exception.getMessage());
+        log.error("error: {}", exception);
+
+        return new RedirectView(request.getHeader("Referer"));
     }
 }
